elf: Move Elf64Phdr to separate module
In the effort of splitting up the ELF parser code, move the Elf64Phdr
implementation to its own module.
